Restore something I commented out previously.
This was something I commented out because I couldn't figure out how to deal with it.  It was simply the wrong version of link() being called.  There is a Link() in the imgui-node-editor api, and this is was being icalled erroniously.  We really needed the constructor of turnkey::types::Link().