Working "pass back heap address" for Save.
this is the first time I've made a function like this, so I was really sussing it out.

I am not impressed with using a pointer-to-pointer as a function argument.
I think I am going to rejigger this so that a size_t* is in the argument, and the function returns a void* or char* since that is less intimidating, I guess.